fix: preload fonts (#4450)
