+++ packages/api-prerendering-service/src/render/types.ts
@@ -22,6 +22,7 @@ export interface RenderHookPlugin extends Plugin {
 export interface RenderApolloState {
     [key: string]: string;
 }
+
 /**
  * @internal
  */
@@ -40,6 +41,7 @@ export interface PeLoaderCacheEntry {
 export interface RenderResult {
     content: string;
     meta: {
+        interceptedRequests: Array<{ type: string; url: string }>;
         apolloState: RenderApolloState;
         cachedData: {
             apolloGraphQl: GraphQLCacheEntry[];
@@ -49,6 +51,7 @@ export interface RenderResult {
         [key: string]: any;
     };
 }
+
 /**
  * @internal
  */
@@ -56,12 +59,14 @@ export interface RenderUrlCallableParams {
     context: Context;
     args: RenderEvent;
 }
+
 /**
  * @internal
  */
 export interface RenderUrlParams extends RenderUrlCallableParams {
     renderUrlFunction?: (url: string, params: RenderUrlCallableParams) => Promise<RenderResult>;
 }
+
 /**
  * @internal
  */
